feat: user-based data isolation in all queries

+++ routes/rental_agreements.py
@@ -1,5 +1,6 @@
 from flask import Blueprint, render_template, request, flash, redirect, url_for
 from db import get_db_con
+from flask_login import current_user
 
 rental_agreements_bp = Blueprint ("rental_agreements", __name__)
 
@@ -22,7 +23,8 @@ def list_rental_agreements():
         FROM rental_agreement ra
         JOIN apartment a ON ra.apartment_id = a.id
         JOIN tenant t ON ra.tenant_id = t.id
-        """).fetchall()
+        WHERE ra.user_id = ?
+        """, (current_user.id,)).fetchall()
     conn.close()
     #function to fetch all apartments and then show it in a table
     return render_template("rental_agreements.html", rental_agreements = rental_agreements)
@@ -34,12 +36,12 @@ def edit_rental_agreement(id):
     if request.method=="GET":
         # Get current agreement
         agreement = db_con.execute("""
-            SELECT * FROM rental_agreement WHERE id = ?
-        """, (id,)).fetchone()
+            SELECT * FROM rental_agreement WHERE id = ? AND user_id = ?
+        """, (id, current_user.id)).fetchone()
 
         # Get options for dropdowns
-        apartment = db_con.execute("SELECT id, name FROM apartment WHERE id = ?", (id,)).fetchone()
-        tenants = db_con.execute("SELECT id, name FROM tenant").fetchall()
+        apartment = db_con.execute("SELECT id, name FROM apartment WHERE id = ? AND user_id", (id, current_user.id)).fetchone()
+        tenants = db_con.execute("SELECT id, name FROM tenant WHERE user_id = ?", (current_user.id)).fetchall()
 
         return render_template("edit_rental_agreement.html", agreement=agreement,
                             apartment=apartment, tenants=tenants)
@@ -49,7 +51,7 @@ def edit_rental_agreement(id):
         start_date = request.form["start_date"]
         end_date = request.form["end_date"]
         rent_amount = request.form["rent_amount"]
-        db_con.execute("UPDATE rental_agreement SET tenant_id = ?, start_date = ?, end_date = ?, rent_amount = ? WHERE id = ?", (tenant_id, start_date, end_date, rent_amount, id))
+        db_con.execute("UPDATE rental_agreement SET tenant_id = ?, start_date = ?, end_date = ?, rent_amount = ? WHERE id = ? AND user_id = ?", (tenant_id, start_date, end_date, rent_amount, id, current_user.id))
         db_con.commit()
         flash("Rental agreement updated successfully.")
         return redirect(url_for("rental_agreements.list_rental_agreements"))
@@ -58,19 +60,19 @@ def edit_rental_agreement(id):
 def delete_rental_agreement(id):
     db_con = get_db_con()
     if request.method=="POST":
-        db_con.execute("DELETE FROM rental_agreement WHERE id = ?",(id,))
+        db_con.execute("DELETE FROM rental_agreement WHERE id = ? AND user_id = ?",(id, current_user.id))
         db_con.commit()
         flash("Successfully deleted")
         return redirect(url_for("rental_agreements.list_rental_agreements"))
     
-    rental_agreement = db_con.execute("SELECT * FROM rental_agreement WHERE id = ?", (id,)).fetchone()
+    rental_agreement = db_con.execute("SELECT * FROM rental_agreement WHERE id = ? AND user_id = ?", (id, current_user.id)).fetchone()
     return render_template("delete_rental_agreement.html", rental_agreement = rental_agreement)
 
 @rental_agreements_bp.route("/rental_agreements/create", methods=["GET", "POST"])
 def create_rental_agreement():
     conn = get_db_con()
     if request.method=="GET":
-        apartments = conn.execute("""SELECT * FROM apartment WHERE id NOT IN (SELECT apartment_id FROM rental_agreement WHERE end_date IS NULL) """).fetchall() #only the ones with active agreement - active = enddate is NULL
+        apartments = conn.execute("""SELECT * FROM apartment WHERE id NOT IN (SELECT apartment_id FROM rental_agreement WHERE end_date IS NULL) AND user_id = ?""",(current_user.id,)).fetchall() #only the ones with active agreement - active = enddate is NULL
         tenants = conn.execute("SELECT id, name FROM tenant").fetchall()
         return render_template("create_rental_agreement.html", apartments = apartments, tenants = tenants)
     else:
@@ -81,9 +83,9 @@ def create_rental_agreement():
         rent_amount = request.form["rent_amount"]
 
         conn.execute("""
-            INSERT INTO rental_agreement (apartment_id, tenant_id, start_date, end_date, rent_amount)
-            VALUES (?, ?, ?, ?, ?)
-        """, (apartment_id, tenant_id, start_date, end_date, rent_amount))
+            INSERT INTO rental_agreement (apartment_id, tenant_id, start_date, end_date, rent_amount, user_id)
+            VALUES (?, ?, ?, ?, ?, ?)
+        """, (apartment_id, tenant_id, start_date, end_date, rent_amount, current_user.id))
         conn.commit()
         return redirect(url_for('rental_agreements.list_rental_agreements'))
